messing with range values for sim settings

+++ Assets/Scripts/Simulation/SimSettings.cs
@@ -32,7 +32,7 @@ public class SimSettings : MonoBehaviour
 
     public void ChangeSliderValue(Slider slider)
     {
-        Debug.Log(slider.value);
+        Debug.Log((int)slider.value);
     }
 
     public void ChangeDropdownValue(TMP_Dropdown dropdown)
